requirements.txt
flask==3.0.0
feedparser==6.0.10
openai==1.12.0
supabase
python-dotenv==1.0.0
requests==2.31.0
